requirements.txt
alabaster==0.7.12
aniso8601==8.0.0
atomicwrites==1.3.0
attrs==19.3.0
Babel==2.7.0
bcrypt==3.1.7
bdc-core==0.3
Cerberus==1.3.1
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
Click==7.0
coverage==4.4.2
docutils==0.15.2
ephem==3.7.7.0
Flask==1.0.3
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-RESTful==0.3.6
flask-restplus==0.12.1
Flask-Script==2.0.6
idna==2.7
imagesize==1.1.0
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==3.1.1
lunardate==0.2.0
MarkupSafe==1.1.1
marshmallow==2.15.1
more-itertools==7.2.0
mysql-connector-python==8.0.18
packaging==19.2
pluggy==0.13.0
protobuf==3.10.0
py==1.8.0
pyCalverter==1.6.1
pycparser==2.19
Pygments==2.4.2
PyJWT==1.7.1
pyparsing==2.4.5
pyrsistent==0.15.5
pytest==5.2.2
pytest-cov==2.8.1
python-dateutil==2.8.1
pytz==2017.3
requests==2.20.0
six==1.11.0
snowballstemmer==2.0.0
Sphinx==2.2.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
urllib3==1.24.3
wcwidth==0.1.7
Werkzeug==0.15.3
workalendar==2.3.1
zipp==0.6.0
